refactor: add prettier and eslint
