+++ src/Components/Search/index.jsx
@@ -1,0 +1,82 @@
+import React from 'react'
+import { fade, makeStyles } from '@material-ui/core/styles'
+import SearchIcon from '@material-ui/icons/Search'
+import InputBase from '@material-ui/core/InputBase'
+import Button from '@material-ui/core/Button'
+
+const useStyles = makeStyles(theme => ({
+  search: {
+    position: 'relative',
+    borderRadius: theme.shape.borderRadius,
+    backgroundColor: fade(theme.palette.common.white, 0.15),
+    '&:hover': {
+      backgroundColor: fade(theme.palette.common.white, 0.25)
+    },
+    marginRight: theme.spacing(1),
+    marginLeft: 0,
+    width: '100%',
+    [theme.breakpoints.up('sm')]: {
+      marginLeft: theme.spacing(3)
+    }
+  },
+  searchIcon: {
+    padding: theme.spacing(0, 2),
+    height: '100%',
+    position: 'absolute',
+    pointerEvents: 'none',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center'
+  },
+  inputRoot: {
+    color: 'inherit'
+  },
+  inputInput: {
+    padding: theme.spacing(1, 1, 1, 0),
+    // vertical padding + font size from searchIcon
+    paddingLeft: `calc(1em + ${theme.spacing(4)}px)`,
+    transition: theme.transitions.create('width'),
+    width: '100%',
+    height: '1.6em'
+  }
+}))
+
+export default function Search(props) {
+  const classes = useStyles()
+
+  let { bookSearchOnChange, inputSearchValue } = props
+
+  const handleChange = event => {
+    inputSearchValue = event.target.value
+  }
+
+  return (
+    <>
+      <div className={classes.search}>
+        <div className={classes.searchIcon}>
+          <SearchIcon />
+        </div>
+        <InputBase
+          placeholder="Busque livros pelo título, autor ou ISBN"
+          fullWidth={true}
+          classes={{
+            root: classes.inputRoot,
+            input: classes.inputInput
+          }}
+          inputProps={{ 'aria-label': 'search' }}
+          id={inputSearchValue}
+          onChange={handleChange}
+        />
+      </div>
+      <Button
+        variant="contained"
+        color="default"
+        size="large"
+        disableElevation
+        onClick={() => bookSearchOnChange(inputSearchValue)}
+      >
+        Buscar
+      </Button>
+    </>
+  )
+}
